Continue logging.

+++ src/logger.rs
@@ -6,12 +6,7 @@ pub struct Logger {}
 // could deduplicate with a macro, but not gonna bother f'now
 // lets structure our logger so that we can grep our logs more easily
 impl Logger {
-    pub fn info(
-        wal_number: Option<usize>,
-        table_name: Option<TableName>,
-        tag: &str,
-        message: &str,
-    ) {
+    pub fn info(wal_number: Option<u64>, table_name: Option<TableName>, tag: &str, message: &str) {
         info!(
             "wal:{} table:{} tag:{} message:{}",
             wal_number
@@ -26,12 +21,7 @@ impl Logger {
     }
 
     #[allow(dead_code)]
-    pub fn error(
-        wal_number: Option<usize>,
-        table_name: Option<TableName>,
-        tag: &str,
-        message: &str,
-    ) {
+    pub fn error(wal_number: Option<u64>, table_name: Option<TableName>, tag: &str, message: &str) {
         error!(
             "wal:{} table:{} tag:{} message:{}",
             wal_number
@@ -46,12 +36,7 @@ impl Logger {
     }
 
     #[allow(dead_code)]
-    pub fn debug(
-        wal_number: Option<usize>,
-        table_name: Option<TableName>,
-        tag: &str,
-        message: &str,
-    ) {
+    pub fn debug(wal_number: Option<u64>, table_name: Option<TableName>, tag: &str, message: &str) {
         debug!(
             "wal:{} table:{} tag:{} message:{}",
             wal_number
@@ -67,7 +52,7 @@ impl Logger {
 
     #[allow(dead_code)]
     pub fn structured_panic(
-        wal_number: Option<usize>,
+        wal_number: Option<u64>,
         table_name: Option<TableName>,
         tag: &str,
         message: &str,
